+++ input.csv
@@ -2,4 +2,4 @@
 350000, 10
 400000, 11
 1000000, -20
-1500000', -20
+1500000, -20
